# Fully port BarrierBeforeFinalMeasurements to rust