--- a/crates/accelerate/src/lib.rs
+++ b/crates/accelerate/src/lib.rs
@@ -14,6 +14,7 @@ use std::env;
 
 use pyo3::import_exception;
 
+pub mod barrier_before_final_measurement;
 pub mod check_map;
 pub mod circuit_library;
 pub mod commutation_analysis;