--- a/qiskit/transpiler/passes/layout/disjoint_utils.py
+++ b/qiskit/transpiler/passes/layout/disjoint_utils.py
@@ -107,7 +107,7 @@ def split_barriers(dag: DAGCircuit):
         num_qubits = len(node.qargs)
         if num_qubits == 1:
             continue
-        if node.op.label:
+        if node.label:
             barrier_uuid = f"{node.op.label}_uuid={uuid.uuid4()}"
         else:
             barrier_uuid = f"_none_uuid={uuid.uuid4()}"
@@ -125,28 +125,30 @@ def split_barriers(dag: DAGCircuit):
 def combine_barriers(dag: DAGCircuit, retain_uuid: bool = True):
     """Mutate input dag to combine barriers with UUID labels into a single barrier."""
     qubit_indices = {bit: index for index, bit in enumerate(dag.qubits)}
-    uuid_map: dict[uuid.UUID, DAGOpNode] = {}
+    uuid_map: dict[str, DAGOpNode] = {}
     for node in dag.op_nodes(Barrier):
-        if node.op.label:
-            if "_uuid=" in node.op.label:
-                barrier_uuid = node.op.label
+        if node.label:
+            if "_uuid=" in node.label:
+                barrier_uuid = node.label
             else:
                 continue
             if barrier_uuid in uuid_map:
                 other_node = uuid_map[barrier_uuid]
                 num_qubits = len(other_node.qargs) + len(node.qargs)
-                new_op = Barrier(num_qubits, label=barrier_uuid)
+                if not retain_uuid:
+                    if isinstance(node.label, str) and node.label.startswith("_none_uuid="):
+                        label = None
+                    elif isinstance(node.label, str) and "_uuid=" in node.label:
+                        label = "_uuid=".join(node.label.split("_uuid=")[:-1])
+                    else:
+                        label = barrier_uuid
+                else:
+                    label = barrier_uuid
+                new_op = Barrier(num_qubits, label=label)
                 new_node = dag.replace_block_with_op([node, other_node], new_op, qubit_indices)
                 uuid_map[barrier_uuid] = new_node
             else:
                 uuid_map[barrier_uuid] = node
-    if not retain_uuid:
-        for node in dag.op_nodes(Barrier):
-            if isinstance(node.op.label, str) and node.op.label.startswith("_none_uuid="):
-                node.op.label = None
-            elif isinstance(node.op.label, str) and "_uuid=" in node.op.label:
-                original_label = "_uuid=".join(node.op.label.split("_uuid=")[:-1])
-                node.op.label = original_label
 
 
 def require_layout_isolated_to_component(