--- a/qiskit/transpiler/passes/utils/barrier_before_final_measurements.py
+++ b/qiskit/transpiler/passes/utils/barrier_before_final_measurements.py
@@ -17,6 +17,7 @@ from qiskit.circuit.barrier import Barrier
 from qiskit.transpiler.basepasses import TransformationPass
 from qiskit.dagcircuit import DAGCircuit, DAGOpNode
 from .merge_adjacent_barriers import MergeAdjacentBarriers
+from qiskit._accelerate.barrier_before_final_measurement import barrier_before_final_measurements
 
 
 class BarrierBeforeFinalMeasurements(TransformationPass):
@@ -33,60 +34,7 @@ class BarrierBeforeFinalMeasurements(TransformationPass):
 
     def run(self, dag):
         """Run the BarrierBeforeFinalMeasurements pass on `dag`."""
-        # Collect DAG nodes which are followed only by barriers or other measures.
-        final_op_types = ["measure", "barrier"]
-        final_ops = []
-        for candidate_node in dag.named_nodes(*final_op_types):
-            is_final_op = True
-
-            for _, child_successors in dag.bfs_successors(candidate_node):
-
-                if any(
-                    isinstance(suc, DAGOpNode) and suc.name not in final_op_types
-                    for suc in child_successors
-                ):
-                    is_final_op = False
-                    break
-
-            if is_final_op:
-                final_ops.append(candidate_node)
-
-        if not final_ops:
-            return dag
-
-        # Create a layer with the barrier and add both bits and registers from the original dag.
-        barrier_layer = DAGCircuit()
-        barrier_layer.add_qubits(dag.qubits)
-        for qreg in dag.qregs.values():
-            barrier_layer.add_qreg(qreg)
-        barrier_layer.add_clbits(dag.clbits)
-        for creg in dag.cregs.values():
-            barrier_layer.add_creg(creg)
-
-        # Add a barrier across all qubits so swap mapper doesn't add a swap
-        # from an unmeasured qubit after a measure.
-        final_qubits = dag.qubits
-
-        barrier_layer.apply_operation_back(
-            Barrier(len(final_qubits), label=self.label), final_qubits, (), check=False
-        )
-
-        # Preserve order of final ops collected earlier from the original DAG.
-        ordered_final_nodes = [
-            node for node in dag.topological_op_nodes() if node in set(final_ops)
-        ]
-
-        # Move final ops to the new layer and append the new layer to the DAG.
-        for final_node in ordered_final_nodes:
-            barrier_layer.apply_operation_back(
-                final_node.op, final_node.qargs, final_node.cargs, check=False
-            )
-
-        for final_op in final_ops:
-            dag.remove_op_node(final_op)
-
-        dag.compose(barrier_layer)
-
+        barrier_before_final_measurements(dag, self.label)
         if self.label is None:
             # Merge the new barrier into any other barriers
             adjacent_pass = MergeAdjacentBarriers()