@mtreinish 2024-09-24
I ran the dedicated asv microbenchmarks on this and it yielded:

```
Benchmarks that have improved:

| Change   | Before [1344cddb] <stash~1>   | After [f62a6d79] <port-barrier-before-final-measure>   |   Ratio | Benchmark (Parameter)                                                  |
|----------|-------------------------------|--------------------------------------------------------|---------|------------------------------------------------------------------------|
| -        | 17.1±0.1ms                    | 12.3±0.1ms                                             |    0.72 | passes.PassBenchmarks.time_barrier_before_final_measurements(14, 1024) |
| -        | 7.42±0.03ms                   | 5.12±0.1ms                                             |    0.69 | passes.PassBenchmarks.time_barrier_before_final_measurements(5, 1024)  |
| -        | 29.7±0.2ms                    | 8.93±0.2ms                                             |    0.3  | passes.PassBenchmarks.time_barrier_before_final_measurements(20, 1024) |

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.
```